src/requirements.txt
matplotlib==3.10.3
mplcursors==0.6
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
Pillow==11.3.0
plotly==6.2.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
scipy==1.16.0
soundfile==0.13.1
